+++ Assets/ApiClient/Core/Helpers/CachedRequestHandling.cs
@@ -18,7 +18,7 @@ namespace ApiClientLib.Helpers
             Ref<UnityWebRequest> requestReference, RequestMethod method, int timeout,
             IReadOnlyDictionary<string, string> persistentHeaders, IReadOnlyDictionary<string, string> customHeaders,
             IReadOnlyDictionary<string, string> queryParams, CancellationToken ct, int instanceID,
-            Action<ApiEventData> onRequestCompleted, string fileExtension)
+            Action<ApiEventData> onRequestCompleted, Action<int> onRequestStatusCodeResolved, string fileExtension)
         {
             var startTime = 0L;
             Logging.StartLogTimer(ref startTime);
@@ -33,12 +33,16 @@ namespace ApiClientLib.Helpers
                 var key = ComputeHash(cleanUrl);
                 var (result, wasCached, statusCode) = await cacheManager.GetOrUpdateAsync(key, fileExtension, cleanUrl,
                     cacheDays, downloadToBytes, deserializeFromPath, ct);
+                var finalStatusCode = wasCached ? -1 : (int)(requestReference.Value?.responseCode ?? statusCode);
+                onRequestStatusCodeResolved?.Invoke(finalStatusCode);
                 Logging.ExecuteLogTrigger(false, false, wasCached, cleanUrl, method, timeout, persistentHeaders,
                     customHeaders, queryParams, requestReference.Value, startTime, onRequestCompleted, instanceID);
                 return (statusCode, result);
             }
             catch (Exception e)
             {
+                var finalStatusCode = (int)(requestReference.Value?.responseCode ?? 0);
+                onRequestStatusCodeResolved?.Invoke(finalStatusCode);
                 Logging.ExecuteLogTrigger(false, false, false, cleanUrl, method, timeout, persistentHeaders,
                     customHeaders, queryParams, requestReference.Value, startTime, onRequestCompleted, instanceID,
                     ex: e);